# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Media Tools
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface VideoUrlRequest {
  /** URL of the video */
  url: string;
}

export interface VideoFormat {
  format_id: string;
  ext: string;
  resolution: string;
  filesize?: number | null;
  vcodec?: string | null;
  acodec?: string | null;
  note?: string | null;
}

export interface VideoInfoResponse {
  title: string;
  thumbnail?: string | null;
  duration?: number | null;
  uploader?: string | null;
  formats: VideoFormat[];
}

export type DownloadRequestType =
  (typeof DownloadRequestType)[keyof typeof DownloadRequestType];

export const DownloadRequestType = {
  video: "video",
  audio: "audio",
} as const;

export interface DownloadRequest {
  /** URL to download from */
  url: string;
  /** Specific format ID (null = best quality) */
  format_id?: string | null;
  type: DownloadRequestType;
}

export type ClipRequestType =
  (typeof ClipRequestType)[keyof typeof ClipRequestType];

export const ClipRequestType = {
  video: "video",
  audio: "audio",
  mp3: "mp3",
} as const;

export interface ClipRequest {
  /** YouTube video URL */
  url: string;
  /** Start time in HH:MM:SS or MM:SS format */
  start_time: string;
  /** End time in HH:MM:SS or MM:SS format */
  end_time: string;
  /** Video format ID */
  format_id?: string | null;
  type: ClipRequestType;
}

export interface TranscribeRequest {
  /** Audio or video file to transcribe */
  file: Blob;
  /** Language hint (ar, en, etc.) */
  language?: string | null;
}

export interface TranscriptionResponse {
  text: string;
  language?: string | null;
  duration?: number | null;
}

export interface ConvertRequest {
  /** Video file to convert */
  file: Blob;
}

export interface AnimeRecognizeRequest {
  /** Image of anime scene/character */
  image?: Blob | null;
  /** Text description of the anime */
  description?: string | null;
}

export interface SourceLink {
  name: string;
  url: string;
  icon?: string | null;
}

export interface AnimeResult {
  title: string;
  title_ar?: string | null;
  title_en?: string | null;
  episode?: number | null;
  similarity?: number | null;
  from?: number | null;
  to?: number | null;
  thumbnail?: string | null;
  anilist_id?: number | null;
  mal_id?: number | null;
  genres?: string[] | null;
  description?: string | null;
  source_links?: SourceLink[];
}

export type AnimeRecognitionResponseMethod =
  (typeof AnimeRecognitionResponseMethod)[keyof typeof AnimeRecognitionResponseMethod];

export const AnimeRecognitionResponseMethod = {
  image: "image",
  text: "text",
} as const;

export interface AnimeRecognitionResponse {
  results: AnimeResult[];
  method: AnimeRecognitionResponseMethod;
}

export interface PodcastRecognizeRequest {
  /** Cover image of the podcast */
  image?: Blob | null;
  /** Audio clip from the podcast */
  audio?: Blob | null;
}

export interface PodcastResult {
  title: string;
  description?: string | null;
  image?: string | null;
  author?: string | null;
  categories?: string[] | null;
  source_links?: SourceLink[];
}

export type PodcastRecognitionResponseMethod =
  (typeof PodcastRecognitionResponseMethod)[keyof typeof PodcastRecognitionResponseMethod];

export const PodcastRecognitionResponseMethod = {
  image: "image",
  audio: "audio",
} as const;

export interface PodcastRecognitionResponse {
  results: PodcastResult[];
  method: PodcastRecognitionResponseMethod;
  /** Transcription of audio (if audio recognition was used) */
  transcription?: string | null;
}