# lib/api-zod/src/generated/types/clipRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Media Tools
 * OpenAPI spec version: 0.1.0
 */
import type { ClipRequestType } from "./clipRequestType";

export interface ClipRequest {
  /** YouTube video URL */
  url: string;
  /** Start time in HH:MM:SS or MM:SS format */
  start_time: string;
  /** End time in HH:MM:SS or MM:SS format */
  end_time: string;
  /** Video format ID */
  format_id?: string | null;
  type: ClipRequestType;
}