# lib/api-zod/src/generated/types/convertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Media Tools
 * OpenAPI spec version: 0.1.0
 */

export interface ConvertRequest {
  /** Video file to convert */
  file: Blob;
}