# lib/api-zod/src/generated/types/downloadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Media Tools
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadRequestType } from "./downloadRequestType";

export interface DownloadRequest {
  /** URL to download from */
  url: string;
  /** Specific format ID (null = best quality) */
  format_id?: string | null;
  type: DownloadRequestType;
}