# lib/api-zod/src/generated/types/podcastRecognitionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Media Tools
 * OpenAPI spec version: 0.1.0
 */
import type { PodcastRecognitionResponseMethod } from "./podcastRecognitionResponseMethod";
import type { PodcastResult } from "./podcastResult";

export interface PodcastRecognitionResponse {
  results: PodcastResult[];
  method: PodcastRecognitionResponseMethod;
  /** Transcription of audio (if audio recognition was used) */
  transcription?: string | null;
}