# lib/api-zod/src/generated/types/podcastRecognizeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Media Tools
 * OpenAPI spec version: 0.1.0
 */

export interface PodcastRecognizeRequest {
  /** Cover image of the podcast */
  image?: Blob | null;
  /** Audio clip from the podcast */
  audio?: Blob | null;
}