# lib/api-zod/src/generated/types/podcastResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Media Tools
 * OpenAPI spec version: 0.1.0
 */
import type { SourceLink } from "./sourceLink";

export interface PodcastResult {
  title: string;
  description?: string | null;
  image?: string | null;
  author?: string | null;
  categories?: string[] | null;
  source_links?: SourceLink[];
}