# lib/api-zod/src/generated/types/sourceLink.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Media Tools
 * OpenAPI spec version: 0.1.0
 */

export interface SourceLink {
  name: string;
  url: string;
  icon?: string | null;
}